[AD] Implement AD annotations V2 for Kubernetes Pods
This adds a new way of annotating K8s pods with AD annotations. Instead
of having to supply 3 different annotations (`check_names`, `init_configs`,
`instances`), users can now provide AD configs in a single `checks`
annotation as in the example below:

```
ad.datadoghq.com/foobar.checks: |
{
	"apache": {
		"instances": [
			{"apache_status_url":"http://%%host%%/server-status?auto2"}
		]
	}
}
```
